// src/api/generated/models/notificationModel.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Tg-Api 
 * OpenAPI spec version: 1.0.0
 */

export interface NotificationModel {
  id: string;
  notifiable_type: string;
  notifiable_id: string;
  type: string;
  title: string;
  /** @nullable */
  title_pt: string | null;
  /** @nullable */
  title_en: string | null;
  message: string;
  /** @nullable */
  message_pt: string | null;
  /** @nullable */
  message_en: string | null;
  translation_status: string;
  /** @nullable */
  read_at: string | null;
  /** @nullable */
  link: string | null;
  created_at: string;
  updated_at: string;
}